Crate sota_client

This is the client in-vehicle portion of the SOTA project. See the main SOTA Server project and associated architecture document for more information.

Reexports

extern crate std as std;

extern crate hyper;

extern crate rustc_serialize;

extern crate time;

extern crate url;

extern crate crypto;

extern crate toml;

extern crate dbus;

extern crate log;

extern crate env_logger;

use std::prelude::v1::*;

Modules

configuration
Parsing of the configuration file of sota_client.

handler
Implements message handling for the rvi module.

jsonrpc
RVI specific implementation of the jsonrpc protocol

main_loop
Main loop, starting the worker threads and wiring up communication channels between them.

message
Defines types and helper functions for messages being passed within or out of the system.

persistence
Handles caching and storage on disk for in-progress transfers and the assembly and verification of finished transfers

rvi
RVI bindings for Rust.

sota_dbus
Implements the DBus interface to the Software Loading Manager

Macros

try_msg_or!
Try to unwrap or log the provided message and run the third argument

try_or!
Try to unwrap or log the error and run the second argument