Module sota_client::configuration::common

Helper functions and traits for all configuration sections.

Reexports

use std::prelude::v1::*;

use toml;

use std::result;

use std::fmt;

Traits

ConfTreeParser
Trait that provides a interface for parsing a (sub-) tree of the configuration.

ParseTomlValue
Trait that provides a interface for parsing a single toml value.

Functions

format_parser_error
Helper function to format a toml::Parser error message to the format used in this implementation. This is only safe to call if the parser is associated with a real file on disk.

get_optional_key
Parse a optional key, returning None if it can't be found.

get_required_key
Parse a required key, returning a appropriate error message, if the key can't be found in the configuration.

stringify
Helper function to copy anything that implements Display to a String.

Type Definitions

Result
Result type used throughout the configuration parser.