Struct sota_client::configuration::configuration::Configuration

pub struct Configuration {
    pub client: ClientConfiguration,
    pub dbus: DBusConfiguration,
}

Type to encode the full configuration.

Fields

client

The client section of the configuration

dbus

The dbus section of the configuration

Methods

impl Configuration

fn read(path: &str) -> Result<Configuration>

Try to read the configuration from the provided path and parse it into a Configuration object. Returns the parsed Configuration on success or the first error message encountered while reading or parsing the configuration file.

Arguments

path: Path to the location of the configuration file.

fn parse(conf: &str) -> Result<Configuration>

Try to parse the given string to a Configuration.

Arguments

conf: The configuration to parse.

fn default_path() -> String

Try to find the configuration file in different paths. First $XDG_CONFIG_HOME/sota/client.toml is tried, then $HOME/.sota/client.toml returns $PWD/.sota/client.toml if none of the above can be found. The case where this file also doesn't exist should be handled by read.