Struct sota_client::configuration::ClientConfiguration

pub struct ClientConfiguration {
    pub storage_dir: String,
    pub rvi_url: Option<String>,
    pub edge_url: Option<String>,
    pub timeout: Option<i64>,
    pub vin_match: i32,
}

Type to encode allowed keys for the client section of the configuration.

Fields

storage_dir

Directory where chunks and packages will be stored.

rvi_url

The full URL where RVI can be reached.

edge_url

The host:port combination where the client should bind and listen for incoming RVI calls.

timeout

How long to wait for further server messages before the Transfer will be dropped.

vin_match

Index of the RVI service URL, that holds the VIN for this device.