Module sota_client::handler

Implements message handling for the rvi module.

Implements individual message handler for the different messages the client might receive, and the general ServiceHandler that parses messages and delegates them to the individual handlers.

This is a reference implementation for the rvi module, that can later be split out into a crate.

Reexports

use std::prelude::v1::*;

use std::sync::Mutex;

use std::collections::HashMap;

use message::{BackendServices, PackageId, Notification};

use persistence::Transfer;

Modules

abort
Handles "Abort Transfer" messages.

chunk
Handles messages transferring single chunks.

finish
Handles "Finish Transfer" messages.

notify
Handles "Notify" messages.

report
Handles "Get All Packages" messages.

service
The main service handler

start
Handles "Start Transfer" messages.

Structs

AbortParams
Type for "Abort Transfer" messages. The package transfer to abort

ChunkParams
Type for messages transferring single chunks.

FinishParams
Type for "Finish Transfer" messages.

NotifyParams
Type for "Notify" messages.

ReportParams
Type for "Get All Packages" messages.

ServiceHandler
Type that encodes a single service handler.

StartParams
Type for "Start Transfer" messages.

Traits

HandleMessageParams
Trait that every message handler needs to implement.

Type Definitions

Transfers
Type alias to hide the internal HashMap, that is used to store Transfers.