Module sota_client::handler::service

The main service handler

Parses incoming messages and delegates them to the appropriate individual message handlers, passing on the results to the main_loop

Reexports

use std::prelude::v1::*;

use jsonrpc;

use jsonrpc::{OkResponse, ErrResponse};

use std::io::{Read, Write};

use std::ops::DerefMut;

use std::sync::{Arc, Mutex};

use std::sync::mpsc::Sender;

use std::thread::sleep_ms;

use time;

use hyper::server::{Handler, Request, Response};

use rustc_serialize::{json, Decodable};

use rustc_serialize::json::Json;

use rvi::{Message, RVIHandler, Service};

use message::{BackendServices, LocalServices, Notification};

use handler::{NotifyParams, StartParams, ChunkParams, FinishParams};

use handler::{ReportParams, AbortParams, HandleMessageParams, Transfers};

use configuration::Configuration;

Structs

ServiceHandler
Type that encodes a single service handler.