Struct sota_client::handler::service::ServiceHandler

pub struct ServiceHandler {
    rvi_url: String,
    sender: Mutex<Sender<Notification>>,
    services: Mutex<BackendServices>,
    transfers: Arc<Mutex<Transfers>>,
    conf: Configuration,
    vin: String,
}

Type that encodes a single service handler.

Holds the necessary state, like in-progress transfers, that are needed for handling incoming messages and sending replies to RVI. Needs to be thread safe as hyper handles requests asynchronously.

Fields

rvi_url

The full URL, where RVI can be reached.

sender

A Sender that connects the handlers with the main_loop.

services

The service URLs that the SOTA server advertised.

transfers

The currently in-progress Transfers.

conf

The full Configuration of sota_client.

vin

The VIN of this device, as returned by RVI.

Methods

impl ServiceHandler

fn new(transfers: Arc<Mutex<Transfers>>, sender: Sender<Notification>, url: String, c: Configuration) -> ServiceHandler

Create a new ServiceHandler.

Arguments

transfers: A Transfers object to store the in-progress Transfers.

sender: A Sender to call back into the main_loop.

url: The full URL, where RVI can be reached.

c: The full Configuration of sota_client.

fn start_timer(transfers: &Mutex<Transfers>, timeout: i64)

Starts a infinite loop to expire timed out transfers. Checks once a second for timed out transfers.

Arguments

transfers: Pointer to a Transfers object, that stores the transfers to be checked for expired timeouts.

timeout: The timeout in seconds.

fn push_notify(&self, m: Notification)

Helper function to send a Notification to the main_loop.

Arguments

m: Notification to send.

fn handle_message_params<D>(&self, message: &str) -> Option<Result<OkResponse<i32>, ErrResponse>> where D: Decodable + HandleMessageParams

Create a message handler D, and let it process the message. If it returns a Notification, forward it to the main_loop. Returns a jsonrpc response indicating success or failure.

Arguments

message: The message, that should be handled.

fn handle_message(&self, message: &str) -> Result<OkResponse<i32>, ErrResponse>

Try to parse the type of a message and forward it to the appropriate message handler. Returns the result of the message handling or a jsonrpc result indicating a parser error.

Needs to be extended to support new services.

Arguments

message: The message that will be parsed.