Struct sota_client::jsonrpc::OkResponse

pub struct OkResponse<T> {
    pub jsonrpc: String,
    pub id: u64,
    pub result: Option<T>,
}

Response to a jsonrpc call, indicating a successful method call.

Fields

jsonrpc

The version of jsonrpc to use, has to be set to "2.0".

id

The identifier of the jsonrpc call this response belongs to. Only unsigned numbers are accepted

result

The result of the method call, if any.

Methods

impl<T> OkResponse<T>

fn new(id: u64, result: Option<T>) -> OkResponse<T>

Returns a new OkResponse

Arguments

id: The identifier of the jsonrpc call the returned response belongs to.

result: The result of the method call, if any.