Module sota_client::main_loop

Main loop, starting the worker threads and wiring up communication channels between them.

Reexports

use std::prelude::v1::*;

use std::sync::mpsc::channel;

use std::thread;

use std::sync::{Arc, Mutex};

use std::collections::HashMap;

use std::ops::Deref;

use rvi;

use handler::ServiceHandler;

use message::{InitiateParams, BackendServices, PackageId};

use message::{Notification, ServerPackageReport, LocalServices, ServerReport};

use configuration::Configuration;

use persistence::Transfer;

use sota_dbus;

Functions

start
Main loop, starting the worker threads and wiring up communication channels between them.