Function sota_client::main_loop::start

pub fn start(conf: &Configuration, rvi_url: String, edge_url: String)

Main loop, starting the worker threads and wiring up communication channels between them.

Arguments

conf: A pointer to a Configuration object see the documentation of the configuration crate.

rvi_url: The URL, where RVI can be found, with the protocol.

edge_url: The host:port combination where the client should bind and listen for incoming RVI calls.