Module sota_client::message

Defines types and helper functions for messages being passed within or out of the system.

Reexports

use std::prelude::v1::*;

use std::prelude::v1::*;

use std::vec::Vec;

use dbus::{Message, MessageItem, FromMessageItem, Error};

use super::package_id::PackageId;

use super::server::BackendServices;

use std::prelude::v1::*;

use super::package_id::PackageId;

use super::client::PackageReport;

use std::prelude::v1::*;

use std::fmt;

use dbus::{FromMessageItem, MessageItem};

use std::ops::Deref;

Modules

client
Various messages and helper functions for them.

initiate
Helper functions for the "Initiate Download" message, that gets sent to the server.

package_id
Helper functions for working with packages.

server
Translation layer for the SOTA server.

Structs

BackendServices
Encodes the service URLs, that the server provides.

ChunkReceived
Encodes the "Chunk Received" message, indicating that a chunk was successfully transferred.

InitiateParams
Encodes the parameters needed for the "Initiate Download" message.

LocalServices
Encodes the list of service URLs the client registered.

PackageId
Encodes a package, defined through a name and version.

PackageReport
Encodes a installation report for a single package.

PackageSum
Encodes a package/checksum pair.

ServerPackageReport
Encodes a package installation report, as required by the SOTA server.

ServerReport
Encodes a installed packages report, as required by the SOTA server.

UserMessage
Encodes the full message, that is sent to indicate updates and to provide the callback URLs to the server.

UserPackage
Encodes the package/size pair, that is sent by the server to notify the client of new updates.

Enums

Notification
Enumerates the different notification types, that are sent to the main_loop.

Traits

ParsePackageReport
Provides an interface to parse a PackageReport