Module sota_client::message::client

Various messages and helper functions for them.

Reexports

use std::prelude::v1::*;

use std::vec::Vec;

use dbus::{Message, MessageItem, FromMessageItem, Error};

use super::package_id::PackageId;

use super::server::BackendServices;

Structs

PackageReport
Encodes a installation report for a single package.

UserMessage
Encodes the full message, that is sent to indicate updates and to provide the callback URLs to the server.

UserPackage
Encodes the package/size pair, that is sent by the server to notify the client of new updates.

Enums

Notification
Enumerates the different notification types, that are sent to the main_loop.

Traits

ParsePackageReport
Provides an interface to parse a PackageReport