Enum sota_client::message::client::Notification

pub enum Notification {
    Notify(UserMessage),
    Initiate(PackageId),
    Report,
    Finish(PackageId),
}

Enumerates the different notification types, that are sent to the main_loop.

Variants

Notify

Sent when new updates are available.

Initiate

Sent when the user wants to update a package.

Report

Sent when the server requested a list of installed packages.

Finish

Sent when a transfer is completed and ready to be installed.