Struct sota_client::message::client::PackageReport

pub struct PackageReport {
    pub package: PackageId,
    pub status: bool,
    pub description: String,
}

Encodes a installation report for a single package.

Fields

package

The package that was installed.

status

Boolean to indicate success or failure of the installation.

description

A short description of the result of the installation.