Struct sota_client::message::ServerPackageReport

pub struct ServerPackageReport {
    pub package: PackageId,
    pub status: bool,
    pub description: String,
    pub vin: String,
}

Encodes a package installation report, as required by the SOTA server.

Fields

package

The package that got installed.

status

Boolean to indicate success or failure.

description

A short description of the result of the installation request.

vin

The VIN of this device.

Methods

impl ServerPackageReport

fn new(r: PackageReport, v: String) -> ServerPackageReport

Create a new ServerPackageReport from a PackageReport.

Arguments

r: The PackageReport to translate for the server.

v: The VIN of this device.