Struct sota_client::persistence::Transfer

pub struct Transfer {
    pub package: PackageId,
    pub checksum: String,
    pub transferred_chunks: Vec<u64>,
    pub prefix_dir: String,
    pub last_chunk_received: i64,
}

Type for storing the metadata of a in-progress transfer, which is defined as one package. Will clear out the chunks on disk when freed.

Fields

package

PackageId of this transfer.

checksum

SHA1 checksum of the fully assembled package.

transferred_chunks

Vector of transferred chunks.

prefix_dir

Path to the directory, where chunks will be cached and finished packages will be stored.

last_chunk_received

Timestamp, when the last chunk was received. Given as a unix epoch timestamp.

Methods

impl Transfer

fn new(prefix: String, package: PackageId, checksum: String) -> Transfer

Return a new Transfer

Arguments

prefix: Path where transferred chunks and assembled package will be stored.

package: PackageId of this transfer.

checksum: SHA1 checksum of the fully assembled package.

fn write_chunk(&mut self, msg: &str, index: u64) -> bool

Write a transferred chunk to disk. Returns false and logs an error if something goes wrong.

Arguments

msg: Base64 encoded data of this chunk.

index: Index of this chunk

fn assemble_package(&self) -> bool

Assemble the transferred chunks to a package and verify it with the provided checksum. Returns false and prints a error message if either the package can't be assembled or the checksum doesn't match.

fn assemble_chunks(&self) -> Result<(), String>

Collect all chunks and concatenate them into one file. Returns a String with a error message, should something go wrong.

fn copy_chunk(&self, path: &PathBuf, index: u64, file: &mut File) -> Result<(), String>

Read a chunk file file and append it to a package file. Returns a String with a error message should something go wrong.

Arguments

path: Pointer to a PathBuf where the chunks are cached.

index: Index of the chunk to append.

file: Pointer to a File where the chunk should be appended. Should be created with OpenOptions and the append only option. See the documentation for OpenOptions, File, and the implementation of assemble_chunks for details.

fn checksum(&self) -> bool

Verify the checksum of this transfer. Assumes the package was already assembled. Prints a error message showing the mismatched checksums and returns false on errors.

fn get_chunk_path(&self, index: u64) -> Result<PathBuf, String>

Get the full path for the specified chunk index. Returns a PathBuf on success or a String on errors detailing what went wrong.

Arguments

index: The index for which the path should be constructed

fn get_package_path(&self) -> Result<PathBuf, String>

Get the full path for the package of this Transfer. Returns a PathBuf on success or a String on errors detailing what went wrong.

fn get_chunk_dir(&self) -> Result<PathBuf, String>

Get the directory, where this Transfer caches chunks. Returns a PathBuf on success or a String on errors detailing what went wrong.

fn get_package_dir(&self) -> Result<PathBuf, String>

Get the directory, where this Transfer stores the assembled package. Returns a PathBuf on success or a String on errors detailing what went wrong.