Module sota_client::rvi

RVI bindings for Rust.

RVI - Remote Vehicle Interaction - is the next generation of connected vehicle services. Based on the discussions inside and outside the Automotive Grade Linux expert group.

This module implements Rust bindings to simplify the interaction with it.

It is intended to be split out into a separate crate at some point in the future.

Reexports

use std::prelude::v1::*;

Modules

edge
Implements the RVI facing webservice.

handler
Trait definition of RVIHandler

message
Wrappers for messages exchanged with RVI.

send
Helper functions for sending messages to RVI.

Structs

Message
A generic incoming message.

Service
Encodes a registered service that this device provides.

ServiceEdge
Encodes the service edge of the webservice.

Traits

RVIHandler
Provides a interface to register services.

Functions

send
Send a object to RVI. Either returns the full response from RVI or a error message.

send_message
Prepare a message and send it to RVI. Returns the full response from RVI on success or a error message on failure.