Function sota_client::rvi::send_message

pub fn send_message<E: Encodable>(url: &str, b: E, addr: &str) -> Result<String, String>

Prepare a message and send it to RVI. Returns the full response from RVI on success or a error message on failure.

This wraps the provided object into a proper RVI message and encodes it to json. You usually should call this function.

NOTE: This currently implements a workaround for RVI, that will get fixed in the upcoming RVI version 0.5.0, which will break this current implementation. For the new protocol you don't have to wrap the params in a one element Vector any more.

Arguments

url: The full URL where RVI can be reached.

b: The object to wrap into a RVI Message, encode and send to RVI.

addr: The full RVI address (service URL) where this message should be sent to.