Struct sota_client::rvi::ServiceEdge

pub struct ServiceEdge {
    rvi_url: String,
    edge_url: String,
    sender: Sender<Vec<Service>>,
}

Encodes the service edge of the webservice.

Fields

rvi_url

The full URL where RVI can be reached.

edge_url

The host:port combination where the edge should bind and listen for incoming RVI messages.

sender

A sender to communicate back the service URLs.

Methods

impl ServiceEdge

fn new(r: String, e: String, s: Sender<Vec<Service>>) -> ServiceEdge

Create a new service edge.

Arguments

r: The full URL where RVI can be reached.

e: The host:port combination where the edge should bind.

s: A sender to communicate back the service URLs.

fn register_service(&self, s: &str) -> String

Register a service. Returns the full service URL as provided by RVI. Panics if the registration in RVI failed. This can be handled by starting the RVI edge in a separate thread.

Arguments

s: The service to register. Will get prepended with the device identifier by RVI.

fn register(&self, svcs: &Vec<&str>) -> Vec<Service>

Helper function to register multiple services. Returns a Vector of Services.

Arguments

svcs: Pointer to a Vector of service strings, that should be registered.

fn start<H: 'static>(&self, h: H, s: Vec<&str>) where H: Handler + RVIHandler

Starts the service edge.

It binds on the provided host:port combination, registers all services and then waits for incoming RVI messages. On incoming messages it forks another thread and passes the message to the provided Handler. For details about how to implement a Handler see the hyper documentation and the reference implementation.

Panics if it can't reach or register in RVI.

Arguments

h: The Handler all messages are passed to.

s: A Vector of service strings to register in RVI.