Module sota_client::sota_dbus

Implements the DBus interface to the Software Loading Manager

Also see the documentation of the Rust dbus bindings.

Reexports

use std::prelude::v1::*;

Modules

receiver
Receiving side of the DBus interface.

sender
Sending side of the DBus interface.

Structs

Receiver
Encodes the state that is needed to accept incoming DBus messages.

Functions

request_install
Ask the Software Loading Manager to isntall a package. Will block until the installation finished or the timeout is reached.

request_report
Request a full report from the Software Loading Manager. Will block until the list of all installed packages is received or the timeout is reached.

send_notify
Foward a "Notify" message to DBus.