Module sota_client::sota_dbus::sender

Sending side of the DBus interface.

Reexports

use std::prelude::v1::*;

use std::convert::From;

use std::borrow::Cow;

use dbus::{Connection, BusType, MessageItem, Message, FromMessageItem};

use configuration::DBusConfiguration;

use message::{UserPackage, PackageId, PackageReport};

use message::ParsePackageReport;

Functions

parse_package_list
Parses an incoming DBus message to a Vector of PackageIds. Ignores unparsable entries, thus an empty Vector might indicate a parser error.

request_install
Ask the Software Loading Manager to isntall a package. Will block until the installation finished or the timeout is reached.

request_report
Request a full report from the Software Loading Manager. Will block until the list of all installed packages is received or the timeout is reached.

send_notify
Foward a "Notify" message to DBus.