Function sota_client::sota_dbus::sender::request_report

pub fn request_report(config: &DBusConfiguration) -> Vec<PackageId>

Request a full report from the Software Loading Manager. Will block until the list of all installed packages is received or the timeout is reached.

Arguments

config: The configuration of the DBus interface.