Struct sota_client::sota_dbus::Receiver

pub struct Receiver {
    config: DBusConfiguration,
    sender: Sender<Notification>,
}

Encodes the state that is needed to accept incoming DBus messages.

Fields

config

The configuration for the DBus interface.

sender

A sender to forward incoming messages.

Methods

impl Receiver

fn new(c: DBusConfiguration, s: Sender<Notification>) -> Receiver

Create a new Receiver.

Arguments

c: The configuration for the DBus interface.

s: A sender to forward incoming messages.

fn start(&self)

Start the listener. It will register in DBus according to the configuration, wait for incoming messages and forward them via the internal Sender.

fn handle_initiate(&self, msg: &mut Message) -> MethodResult

Handles incoming "Initiate Download" messages.

Parses the message and forwards it to the internal Sender.

Arguments

msg: The message to handle.